chore: step 패키지 제거 및 하나로 통일
