+++ src/main/kotlin/calculator/main.kt
@@ -1,0 +1,5 @@
+package calculator
+
+fun main() {
+    calculate("12321 - 3 / 4 * 2")
+}
