+++ src/main/kotlin/carRacing/main.kt
@@ -1,0 +1,8 @@
+package carRacing
+
+import carRacing.model.CarRacing
+
+
+fun main() {
+    CarRacing().start()
+}
