+++ src/main/kotlin/carRacing/model/CarGenerator.kt
@@ -1,0 +1,5 @@
+package carRacing.model
+
+object CarGenerator {
+    fun generate(carNames: List<String>) = carNames.map { Car(carName = it) }
+}
