--- src/main/kotlin/step2/stringCalculator/main.kt
@@ -1,5 +1,0 @@
-package step2.stringCalculator
-
-fun main() {
-    calculate("12321 - 3 / 4 * 2")
-}
