--- src/main/kotlin/step3/carRacing/RaceRule.kt
@@ -1,7 +1,0 @@
-package step3.carRacing
-
-import kotlin.random.Random
-
-object RaceRule {
-    val over4CanGo = Random.nextInt(0, 10) in 4 until 10
-}
