--- src/main/kotlin/step3/carRacing/model/CarGenerator.kt
@@ -1,7 +1,0 @@
-package step3.carRacing.model
-
-class CarGenerator {
-    companion object {
-        fun generate(size: Int) = List<Car>(size) { Car() }
-    }
-}
