--- src/main/kotlin/step3/carRacing/model/CarRacing.kt
@@ -1,17 +1,0 @@
-package step3.carRacing.model
-
-import step3.carRacing.RaceRule
-import step3.carRacing.view.InputView
-import step3.carRacing.view.OutputView
-import step3.carRacing.view.SystemInputReader
-import step3.carRacing.view.SystemPrint
-
-class CarRacing {
-
-    fun start() {
-        val (carCount, tryCount) = InputView.getInputInfo(SystemInputReader())
-        val raceOrganizer = RaceOrganizer(cars = CarGenerator.generate(carCount))
-        raceOrganizer.startRace(tryCount) { RaceRule.over4CanGo }
-        OutputView.renderRace(raceOrganizer.getRaceResult(), SystemPrint)
-    }
-}
