--- src/main/kotlin/step4/carRacingWinner/InputRule.kt
@@ -1,16 +1,0 @@
-package step4.carRacingWinner
-
-object InputRule {
-
-    private fun <T> validate(list: List<T>, condition: (param: T) -> Boolean, errorString: String): List<T> {
-        if (!list.all { condition(it) }) {
-            throw IllegalArgumentException(errorString)
-        }
-        return list
-    }
-
-    val validateCarNames = { carNames: List<String> ->
-        validate(carNames, { it.length < 6 }, "자동차 이름은 5자를 초과할 수 없습니다")
-    }
-
-}
