--- src/main/kotlin/step4/carRacingWinner/main.kt
@@ -1,7 +1,0 @@
-package step4.carRacingWinner
-
-import step4.carRacingWinner.model.CarRacing
-
-fun main() {
-    CarRacing().start()
-}
