--- src/main/kotlin/step4/carRacingWinner/model/CarGenerator.kt
@@ -1,5 +1,0 @@
-package step4.carRacingWinner.model
-
-object CarGenerator {
-    fun generate(carNames: List<String>) = carNames.map { Car(carName = it) }
-}
